fix warnings and small ajustments
