+++ src/utils/formatDate.js
@@ -1,0 +1,10 @@
+const formatDate = timestamp => {
+  const date = new Date(timestamp)
+  const day = date.getDate()
+  const month = date.getMonth() + 1
+  const year = date.getFullYear()
+
+  return `${day}/${month}/${year}`
+}
+
+export default formatDate
